pdf_summary_bot/requirements.txt
streamlit
faiss-cpu
PyPDF2
langchain
langchain-community
langchain-core
langchainhub
langchain-text-splitters
sentence-transformers